Added Animator
Now there is an option for annealing and GD to return a list of Y_s over optimization iteration for animation purposes - I believe there is a bug somewhere in how we model pass through SPN, since the times taken in main.ipynb are way longer than the ones reported on SpeedTest.ipynb. This should be investigated.